add invite entity action + clean up
