+++ src/packages/user/user/invite/entity-action/invite/invite-user-entity-action.ts
@@ -1,10 +1,12 @@
+import { UMB_INVITE_USER_MODAL } from '../../index.js';
 import { UmbEntityActionBase } from '@umbraco-cms/backoffice/entity-action';
 import { UMB_MODAL_MANAGER_CONTEXT } from '@umbraco-cms/backoffice/modal';
 
 export class UmbCreateUserEntityAction extends UmbEntityActionBase<never> {
 	override async execute() {
 		const modalManager = await this.getContext(UMB_MODAL_MANAGER_CONTEXT);
-		debugger;
+		const modalContext = modalManager.open(this, UMB_INVITE_USER_MODAL);
+		await modalContext?.onSubmit();
 	}
 }
