+++ src/packages/user/user/invite/entity-action/invite/manifests.ts
@@ -1,4 +1,4 @@
-import { UMB_USER_ROOT_ENTITY_TYPE } from '../../entity.js';
+import { UMB_USER_ROOT_ENTITY_TYPE } from '../../../entity.js';
 
 import type { ManifestTypes, UmbBackofficeManifestKind } from '@umbraco-cms/backoffice/extension-registry';
